% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Deconvolution}
\alias{Deconvolution}
\title{Deconvolution Class}
\format{
An object of class \code{environment} of length 9.
}
\usage{
Deconvolution
}
\description{
Environment containing methods for the Deconvolution class.
}
\section{Methods}{


\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Assuming x is an object of class Deconvolution
x$lorentzians()  
x$mse()  
x$par_superposition_vec(chemical_shifts)  
x$read_bin(path)  
x$read_json(path)  
x$superposition(chemical_shift)  
x$superposition_vec(chemical_shifts)  
x$write_bin(path)  
x$write_json(path)  
}\if{html}{\out{</div>}}

For more information on the methods, see the Rust documentation at
\url{https://github.com/SombkeMaximilian/metabodecon-rust}.
}

\keyword{datasets}
