% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Spectrum}
\alias{Spectrum}
\title{Spectrum Class}
\format{
An object of class \code{environment} of length 19.
}
\usage{
Spectrum
}
\description{
Environment containing methods for the Spectrum class.
}
\section{Methods}{


\if{html}{\out{<div class="sourceCode R">}}\preformatted{x <- Spectrum$new(chemical_shifts, intensities, signal_boundaries)  
x$chemical_shifts()  
x$frequency()  
x$intensities()  
x$nucleus()  
x$read_bin(path)  
x$read_bruker(path, experiment, processing, signal_boundaries)  
x$read_bruker_set(path, experiment, processing, signal_boundaries)  
x$read_jcampdx(path, signal_boundaries)  
x$read_jcampdx_set(path, signal_boundaries)  
x$read_json(path)  
x$reference_compound()  
x$set_frequency(frequency)  
x$set_nucleus(nucleus)  
x$set_reference_compound(reference)  
x$set_signal_boundaries(signal_boundaries)  
x$signal_boundaries()  
x$write_bin(path)  
x$write_json(path)  
}\if{html}{\out{</div>}}

For more information on the methods, see the Rust documentation at
\url{https://github.com/SombkeMaximilian/metabodecon-rust}.
}

\keyword{datasets}
