% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Deconvoluter}
\alias{Deconvoluter}
\title{Deconvoluter Class}
\format{
An object of class \code{environment} of length 15.
}
\usage{
x <- Deconvoluter$new()  
x$add_ignore_region(start, end)  
x$clear_ignore_regions()  
x$deconvolute_spectra(spectra)  
x$deconvolute_spectrum(spectrum)  
x$fitting_settings()  
x$ignore_regions()  
x$optimize_settings(reference)  
x$par_deconvolute_spectra(spectra)  
x$par_deconvolute_spectrum(spectrum)  
x$selection_settings()  
x$set_analytical_fitter(iterations)  
x$set_moving_average_smoother(iterations, window_size)  
x$set_noise_score_selector(threshold)  
x$smoothing_settings()
}
\description{
Environment containing methods for the Deconvoluter class.
}
\details{
For more information on the methods, see the Rust documentation at
\url{https://github.com/SombkeMaximilian/metabodecon-rust}.
}
\keyword{datasets}
